Move non requirement command classes out of req_command.py
By moving SessionCommandMixin and IndexGroupCommand into their own
module, lazy imports can be easily used.
